+++ gulp/tasks/styles.js
@@ -1,12 +1,13 @@
 'use strict';
 
 var gulp = require( 'gulp' );
-var $ = require( 'gulp-load-plugins' )();
+var plugins = require( 'gulp-load-plugins' )();
 var mqr = require( 'gulp-mq-remove' );
-var pkg = require( '../config' ).pkg;
-var banner = require( '../config' ).banner;
-var config = require( '../config' ).styles;
-var handleErrors = require( '../utils/handleErrors' );
+var config = require( '../config' );
+var configPkg = config.pkg;
+var configBanner = config.banner;
+var configStyles = config.styles;
+var handleErrors = require( '../utils/handle-errors' );
 var browserSync = require( 'browser-sync' );
 
 /**
@@ -14,20 +15,20 @@ var browserSync = require( 'browser-sync' );
  * @returns {PassThrough} A source stream.
  */
 function stylesModern() {
-  return gulp.src( config.cwd + config.src )
-    .pipe( $.sourcemaps.init() )
-    .pipe( $.less( config.settings ) )
+  return gulp.src( configStyles.cwd + configStyles.src )
+    .pipe( plugins.sourcemaps.init() )
+    .pipe( plugins.less( configStyles.settings ) )
     .on( 'error', handleErrors )
-    .pipe( $.autoprefixer( {
+    .pipe( plugins.autoprefixer( {
       browsers: [ 'last 2 version',
                   'not ie <= 8',
                   'android 4',
                   'BlackBerry 7',
                   'BlackBerry 10' ]
     } ) )
-    .pipe( $.header( banner, { pkg: pkg } ) )
-    .pipe( $.sourcemaps.write( '.' ) )
-    .pipe( gulp.dest( config.dest ) )
+    .pipe( plugins.header( configBanner, { pkg: configPkg } ) )
+    .pipe( plugins.sourcemaps.write( '.' ) )
+    .pipe( gulp.dest( configStyles.dest ) )
     .pipe( browserSync.reload( {
       stream: true
     } ) );
@@ -38,22 +39,22 @@ function stylesModern() {
  * @returns {PassThrough} A source stream.
  */
 function stylesIe() {
-  return gulp.src( config.cwd + config.src )
-    .pipe( $.less( config.settings ) )
+  return gulp.src( configStyles.cwd + configStyles.src )
+    .pipe( plugins.less( configStyles.settings ) )
     .on( 'error', handleErrors )
-    .pipe( $.autoprefixer( {
+    .pipe( plugins.autoprefixer( {
       browsers: [ 'ie 7-8' ]
     } ) )
     .pipe( mqr( {
       width: '75em'
     } ) )
     // mqr expands the minified file
-    .pipe( $.cssmin() )
-    .pipe( $.rename( {
+    .pipe( plugins.cssmin() )
+    .pipe( plugins.rename( {
       suffix:  '.ie',
       extname: '.css'
     } ) )
-    .pipe( gulp.dest( config.dest ) )
+    .pipe( gulp.dest( configStyles.dest ) )
     .pipe( browserSync.reload( {
       stream: true
     } ) );
@@ -64,18 +65,18 @@ function stylesIe() {
  * @returns {PassThrough} A source stream.
  */
 function stylesOnDemand() {
-  return gulp.src( config.cwd + '/on-demand/*.less' )
-    .pipe( $.less( config.settings ) )
+  return gulp.src( configStyles.cwd + '/on-demand/*.less' )
+    .pipe( plugins.less( configStyles.settings ) )
     .on( 'error', handleErrors )
-    .pipe( $.autoprefixer( {
+    .pipe( plugins.autoprefixer( {
       browsers: [ 'last 2 version',
                   'ie 7-8',
                   'android 4',
                   'BlackBerry 7',
                   'BlackBerry 10' ]
     } ) )
-    .pipe( $.header( banner, { pkg: pkg } ) )
-    .pipe( gulp.dest( config.dest ) )
+    .pipe( plugins.header( configBanner, { pkg: configPkg } ) )
+    .pipe( gulp.dest( configStyles.dest ) )
     .pipe( browserSync.reload( {
       stream: true
     } ) );