+++ CHANGELOG.md
@@ -35,12 +35,19 @@ Given the `MAJOR.MINOR.PATCH` pattern, here is how we decide to increment:
 - missing publish_eccu requirements
 
 ### Changes
+- Frontend: Added `destroyInitFlag()` method to `atomic-helpers.js`.
+- Frontend: Added `destroy()` method to `Expandables.js` to allow
+  reversing calls to `init()`.
+- Frontend: Added extra small tests to bureau structure page.
 
 ### Removed
 
 - Removed activities-block.html
 
 ### Fixed
+- Frontend: Fixed issue where cloned expandables were not initializing
+  on the bureau structure page.
+- Frontend: Removed `self` references in ContentSlider.
 
 
 ##3.0.0-3.3.17 - 2016-05-20